read code for dense hll

+++ velox/common/hyperloglog/HllUtils.h
@@ -59,8 +59,8 @@ inline uint32_t computeIndex(uint64_t hash, int indexBitLength) {
 /// Returns number of contiguous zeros after 'indexBitLength' bits in the
 /// 'hash'.
 ///
-/// 拿到(去掉 indexBit 之后) Leading zero 的记数.
-inline int numberOfLeadingZeros(uint64_t hash, int indexBitLength) {
+/// 拿到(去掉 indexBit 之后) Leading zero 的记数, 是个 clz 的结果.
+inline int8_t numberOfLeadingZeros(uint64_t hash, int indexBitLength) {
   // Place a 1 in the LSB to preserve the original number of leading zeros if
   // the hash happens to be 0.
   return __builtin_clzl(
